requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.26.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
setuptools==69.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
